2008-01-04  Chris Toshok  <toshok@ximian.com>
        * runtime.cpp: mostly switch all emit_* calls to take doubles,
        * and
        pass doubles around for button x/y.
        (Surface::motion_notify_callback): if we're gtk > 2.12, use
        gdk_event_request_motions (falling back to gdk_window_get_pointer
        if not).  also, we can use the mouse_event_x/y from the event,
        even when it's a hint.

        * runtime.h: switch back to using doubles for everything mouse
        event related.  Change MoonlightEventEmitFunc to take doubles.


svn path=/trunk/moon/; revision=92270

+++ src/runtime.h
@@ -53,7 +53,7 @@ enum RuntimeInitFlags {
 
 class Surface;
 typedef void (* MoonlightFPSReportFunc) (Surface *surface, int nframes, float nsecs, void *user_data);
-typedef void (* MoonlightEventEmitFunc) (UIElement *element, int state, int x, int y);
+typedef void (* MoonlightEventEmitFunc) (UIElement *element, int state, double x, double y);
 
 class Surface : public EventObject {
  public:
@@ -220,8 +220,8 @@ private:
 	int frames;
 
 	int mouse_event_state;
-	int mouse_event_x;
-	int mouse_event_y;
+	double mouse_event_x;
+	double mouse_event_y;
 	
 	// Variables for reporting FPS
 	MoonlightFPSReportFunc fps_report;
